requirements.txt
youtube-transcript-api==0.6.2
pydantic==2.8
python-dotenv>=0.10.0
pinecone-client==5.0.1
tiktoken==0.7.0
pytube==15.0.0
langchain-openai==0.1.22
langchain-community==0.2.12
langchain-pinecone==0.1.3
openai
fastapi